refactor(compaction): extract the added units so oversized files stop growing
This branch added code to files already past the 250 pure-LOC ceiling. Moves compaction execution, summary-request generation, the before-agent-start handler, and the compaction settings surface into modules named for what they own, so index.ts, compaction.ts, speculative.ts, settings-manager.ts, and the compaction suites all return at or below their pre-change size. Pure structural change: no behavior, no assertion, and no test altered.
